[#22] Restructure directories
