+++ src/common/domain/model/NextPage.ts
