+++ src/common/domain/model/Properties.ts
