+++ src/common/domain/service/CommonErrorService.ts
@@ -1,4 +1,4 @@
-import RepositoryError, { RepositoryErrorData } from "../RepositoryError";
+import RepositoryError, { RepositoryErrorData } from "../model/RepositoryError";
 
 export default interface CommonErrorService {
   createRepositoryErrorFrom(e?: RepositoryErrorData): RepositoryError
