+++ src/common/infrastructure/service/CommonErrorServiceImpl.ts
@@ -1,4 +1,4 @@
-import RepositoryError, { RepositoryErrorData } from "src/common/domain/RepositoryError";
+import RepositoryError, { RepositoryErrorData } from "src/common/domain/model/RepositoryError";
 import CommonErrorService from "src/common/domain/service/CommonErrorService";
 
 class CommonErrorServiceImpl implements CommonErrorService {
