+++ src/mother/notice/domain/model/Notice.ts
@@ -1,4 +1,4 @@
-import Id from 'src/common/domain/Id';
+import Id from 'src/common/domain/model/Id';
 
 export default interface Notice {
   id: Id;
