+++ src/mother/notice/domain/service/NoticeService.ts
@@ -1,6 +1,6 @@
-import Id from 'src/common/domain/Id';
-import Page from "src/common/domain/Page";
-import PageRequest from "src/common/domain/PageRequest";
+import Id from 'src/common/domain/model/Id';
+import Page from "src/common/domain/model/Page";
+import PageRequest from "src/common/domain/model/PageRequest";
 import NoticeRequestDto from "../../api/dto/NoticeRequestDto";
 import Notice from "../model/Notice";
