+++ src/mother/notice/infrastructure/repository/NoticeRepositoryImpl.ts
@@ -1,7 +1,7 @@
 import Axios from 'axios';
-import Id from 'src/common/domain/Id';
-import Page from 'src/common/domain/Page';
-import PageRequest from 'src/common/domain/PageRequest';
+import Id from 'src/common/domain/model/Id';
+import Page from 'src/common/domain/model/Page';
+import PageRequest from 'src/common/domain/model/PageRequest';
 import CommonErrorServiceImpl from 'src/common/infrastructure/service/CommonErrorServiceImpl';
 import Notice from '../../domain/model/Notice';
 import NoticeRepository from '../../domain/repository/NoticeRepository'
