+++ src/mother/notice/tests/infrastructure/repository/NoticeRepositoryImpl.unit.test.ts
@@ -2,7 +2,7 @@ import _Axios from 'axios';
 jest.mock('axios')
 
 import { Page } from 'csstype';
-import RepositoryError from 'src/common/domain/RepositoryError';
+import RepositoryError from 'src/common/domain/model/RepositoryError';
 import { getRepositoryErrorFixture } from 'src/common/tests/domain/model/RepositoryError.unit.test';
 import Notice from 'src/mother/notice/domain/model/Notice';
 import { noticeRepository } from 'src/mother/notice/infrastructure/repository/NoticeRepositoryImpl';
